+++ src/FusionEKF.h
@@ -39,8 +39,8 @@ private:
   long previous_timestamp_;
 
   //acceleration noise components
-  float noise_ax;
-  float noise_ay;
+  double noise_ax;
+  double noise_ay;
 
   Eigen::MatrixXd F_;
   Eigen::MatrixXd Q_;
